ml/requirements.txt
# ML / Deep Learning
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
datasets>=2.16.0
sentence-transformers>=2.3.0
accelerate>=0.25.0

# Experiment Tracking & Evaluation
mlflow>=2.10.0
evaluate>=0.4.1
rouge-score>=0.1.2
nltk>=3.8.1

# Document Processing
Pillow>=10.1.0
PyPDF2>=3.0.1
python-docx>=1.1.0
beautifulsoup4>=4.12.2

# Vector Store & Object Storage
qdrant-client>=1.7.0
minio>=7.2.0

# Data Science
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# ONNX Export & Quantization
onnx>=1.15.0
onnxruntime>=1.16.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
